Bug #30520714: MYSQL SERVER RETURNS UNEXPECTED ROWS WHEN USING STORED                FUNCTIONS IN WHERE CLAUSE
When optimizer extracts conditions on constant tables for early evaluation,
it does not include conditions that are expensive to evaluate
which includes conditions involving stored functions. However,
"make_join_select" wrongly concludes that it can remove the entire
"where condition" when the extracted condition evaluates to true because
it involves only const tables (values are available).

Solution:
Added a check for "expensive conditions" before removing the "where_cond"

Reviewed By: Roy Lyseng <roy.lyseng@oracle.com>