Bug#30401004 MAIN.SUBQUERY_MAT_MIXED_TYPES FAILS SPORADICALLY ON GCOV
Sometimes, some EXPLAIN output differed.
First patch, for first problem:
the test has an iteration to test subquery materialization, and another
to test semijoin materialization. In the latter, when semijoin
transformation doesn't apply (subquery is in HAVING), the optimizer chooses
between IN-to-EXISTS and subquery materialization; it is a cost-based choice
which isn't totally repeatable, in spite of ANALYZE TABLE.
Fix: force subquery materialization in this case, using
subquery_materialization_cost_based=off .
I checked all changes in the output. Simply, some IN-to-EXISTS is
replaced with subquery materialization. This changes the result in a
few places, as previously noted in the test it's bug75644 as usual.

Reviewed By : Roy Lyseng <roy.lyseng@oracle.com>

Change-Id: I784761aa93154a36dc4b7428b3b9db67ba8364a2