Bug#30401004 MAIN.SUBQUERY_MAT_MIXED_TYPES FAILS SPORADICALLY ON GCOV
Sometimes, some EXPLAIN output differed.
Second patch, for second problem:
in some cases, duplicates-weedout is chosen, which is correct; but
then the two tables are ordered in a not fully repeatable way (probably
because they're similar in size and cost of access).
So we expect to get plan t1-t2 and we get t2-t1, rarely.
What we want to know, really, is if materialization was chosen or not.
Fix: eliminate all columns of EXPLAIN but the one which tells if
materialization was chosen or not ("SUBQUERY" word, and not
"DEPENDENT SUBQUERY"; "SIMPLE" denotes another strategy).

Reviewed By : Roy Lyseng <roy.lyseng@oracle.com>

Change-Id: Ic2d2599a172084613b5d0b0d2e381e69abce5ebe