Bug#30622327: JSON SCHEMA FUNCTIONS SHOULD CHECK ARGUMENT TYPES AT RESOLVE TIME
Each time JSON_SCHEMA_VALID and JSON_SCHEMA_VALIDATION_REPORT are
called, they check the types of their arguments to see if they are
convertible to JSON. Since the types don't change between invocations,
it suffices to check the types when the functions are resolved.

This patch moves the checks from the function evaluation to the
function resolving (fix_fields()). The checks during evaluation are
replaced with asserts.

Change-Id: I237830acddf36e0c20f9ec1473e3d9209f1816dc