WL#13758 MySQL 5.7: Support linking with 3rd party OpenLDAP and Cyrus SASL
We have recently upgraded to build with OpenSSL 1.1.1 on several
platforms.  This means that the "system" versions of libraries like
LDAP and SASL can no longer be used (they link with other versions of
SSL)

This patch is for 5.7: add cmake options WITH_LDAP and WITH_SASL,
assuming these have been built with the same version of SSL libraries
used for the server.  Look up static (.a) libraries for LDAP, LBER,
and SASL, and link these into appropriate plugins.

We also enable some tests which were previously skipped because of
bad feature tests.

Change-Id: If5deba8ae593b2c666ce12fab6ac7827e41b4f99