Bug#30461911 ASSERTION FAILURE: SRV0SRV.CC:2466:(&(UNDO::DDL_MUTEX))->IS_OWNED()
Issue
=====
The bug was fixed as part of commit 1893011077a9d2b836d88f1b0e2a29198e96487b.
But the commit does not contain a test case.

Fix
===
Adding the testcase as part of this patch. Also extending the other test cases
in the innodb_undo suite to test the test cases with encyption on.

RB: 23494
Reviewed-by: Kevin Lewsis<kevin.lewis@oracle.com>