Bug#30076276 NDB : CLUSTERJ SESSION.FIND() MAPS SOME ERROR CATEGORIES TO NODATAFOUND
When an NDB transaction uses AO_IgnoreError AbortOption during execution,
it essentially means that the transaction will continue executing even
after an operation has failed. On completion of execution, the transaction
will still return 0 indicating success. The individual operations has to be
checked to see if any of them has failed. For this purpose, ClusterJ
maintains a list of operations, that need to be checked for errors, after
the transaction has been executed. But any read operation that is a part of
such transaction is not added to that list.

So, if a transaction having read operations is executed with AO_IgnoreError
option, and if the read operation fails due to a temporary error, the
transaction will still return success. And ClusterJ, as it fails to verify
the read operation for any errors, will interpret it as the row not
existing despite the actual existence of the row.

This patch fixes this issue by adding the read operations to the list
of operations that need to be checked for errors post transaction
execution.

Changes:
Operation.java
 - Added a new isReadOperation() method to the interface

OperationImpl.java
NdbRecordUniqueKeyOperationImpl.java
NdbRecordOperationImpl.java
NdbRecordKeyOperationImpl.java
 - Implemented the newly added isReadOperation() to return true if the
   operation is a read and false otherwise.
 - Read operations are now added to the list of operations that needs
   to be checked for error post transaction execution.

ClusterTransactionImpl.java
 - Updated the error handling in performPostExecuteCallbacks() method.

FindByPrimaryKeyErrorHandlingTest.java
 - Added a new test to verify the error handling.

MgmClient.java
 - Implemented a client that sends command to the managament server by
   connecting to it through sockets. This is used by the newly added
   testcase to provoke an error to verify that the error is being
   handled properly.

AllTests.java
 - Added a new 'enable-debug-tests' options to the class that enables
   all the tests marked with 'DebugTest' annotation. These tests will
   be skipped if the option is disabled.
 - Removed the 'only-run-slow-tests' option as it is not being used
   anymore.

AbstractClusterJTest.java
 - Updated the verifyException() method to add a single line mode
   option to the regex before looking for a match in the Exception
   message.

Change-Id: I221c5c1eff85abc6708de19ef34c2bbe48de2f28