Bug#30330448 UPGRADE FROM 5.7 TO 8.0.17 CAUSES DUPLICATE ENTRIES FOR PARTITIONED TABLE
Problem
-------
During upgrade, while loading the indexes, the clustered indexes named as
GEN_CLUST_INDEX are renamed as PRIMARY as MySQL understands only PRIMARY
indexes.
As part of the upgrade, mysql.innodb_index_stats is updated to add entries
for missing indexes in addition to the entries already existing in 5.7.
Since, the cluster index has been renamed to PRIMARY, an entry for the
same name is not found in the table and hence, a new entry is made which
results in duplicate entries for the same index even if there is no explicit
primary index on the table.

Solution
--------
A cluster index can not have it's name as PRIMARY unless, it was renamed
to PRIMARY earlier during upgrade process. While updating the stats in
innodb_index_stats, ignore the entries with name as PRIMARY and index
type as clustered as this condition can only be true during upgrade.

RB: 23516
Reviewed by: Kevin Lewis (kevin.lewis@oracle.com)