Bug#29115386: CONTRIBUTION: IMPOSSIBLE WHERE FOR A!=A, A<A, A>A
Problem:
If a column is compared to itself in a boolean comparison function like
"a<a", "a>a", "a<>a" and the column is nullable, optimizer does not
detect that these conditions can evaluate to "ALWAYS FALSE" if present
in a WHERE/HAVING clause. For non-nullable columns, optimizer currently
detects such conditions and returns early.

Solution:
While optimizing conditions like above in WHERE/HAVING clause,
for nullable columns, if the eq_cmp_result()
(returns true if the operator would return true on equality, e.g. <=)
returns FALSE, it is treated as an "ALWAYS FALSE"
condition and the server can return early. This is because, when the
value is "NULL", result of the comparison is UNKNOWN and in WHERE/HAVING
clause, UNKNOWN result evaluates to "not true".
According to standard:
    7.12.

    <where clause> ::= WHERE <search condition>

    General Rules
    1) Let T be the result of the preceding <from clause>.
    2) The <search condition> is effectively evaluated for each row of
    T. The result of the <where clause> is a table of those rows of T for
    which the result of the <search condition> is True.

When the value is not NULL and the operator does not allow equality,
as any non-null value cannot be unequal to itself, we can say the result
of comparison will always be false.

The optimization is applicable only to conditions that have the
implicit IS TRUE truth value test. If an explicit truth value
test (such as IS UNKNOWN or IS NOT FALSE, or even IS TRUE) is
specified, the optimization is not applied.

Reviewed By : Dag Wanvik <dag.wanvik@oracle.com>
              Roy Lyseng <roy.lyseng@oracle.com>