Bug#30628330: INSERT WITH JSON_TABLE FAILS DUE TO WARNING INVOKED               WITH ON ERROR CLAUSE
When JSON_TABLE is used in an INSERT statement running in strict mode,
conversion errors handled by the ON ERROR clause could make the INSERT
fail. Since the error is handled by the ON ERROR clause, the statement
should not fail unless ERROR ON ERROR is specified.

Fixed by ignoring warnings when converting values to the target type
if DEFAULT ... ON ERROR or NULL ON ERROR has been specified or is
implied.

Change-Id: Icbd765bb654952f46039b03d116469be6d03a8fb