Bug#30350696 REFERENCE FROM DERIVED TABLE JOIN CONDITION TO PRECEDING TABLE COLUMN FAILS Bug#11748138 ERROR 1054 (42S22): UNKNOWN COLUMN 'T2.I2' IN 'ON CLAUSE'
So far, in an ON condition of a JOIN, only references to columns of
the left and right operands of the JOIN were allowed. We here allow
outer references, i.e. references to columns from tables of outer
query blocks:
select (select 1 from t1 left join t2 on t0.a=3) from t0;
becomes allowed.
Note that outer references in WHERE clause were already allowed, as in:
select (select 1 from t1 left join t2 on true where t0.a=3) from t0;

References to other tables of the same FROM clause as
the JOIN are not outer and remain forbidden, like here:
select * from t0 cross join (t1 left join t2 on t0.a=3)
If you want that type of referencing, you have to use a LATERAL join,
which is allowed as t0.a is now an outer reference:
select * from
  t0 join lateral (select * from t1 left join t2 on t0.a=3);

The behaviours above are compliant with the SQL standard.

Implementation:
(1) When building the Name_resolution_context for the ON condition,
give it an outer context; this is the gist of the implementation.
(2) As we can now have outer refs in ON conditions, any code which
wants to gather outer query's tables used by a subquery must look, not
only at WHERE as before, but to ON conditions too; thus, when we
compute sj_corr_tables, we have to do that. A function
walk_join_list() is introduced to that aim.
That also fixes a bug in a case where an ON condition was created by
derived table merging (see test in join_on_outer_ref.test).
(3) As we now do walk_join_list() to collect table maps used
in ON conditions, for sj_corr_tables, let's also collect, at the same
place, table maps used by lateral derived tables, so we can remove
that collection from fix_tables_after_pullout() where it was a bit
"alien". That's a pure code cleanup.
(4) A follow-up refactoring:
replaced SELECT_LEX::decorrelate_join_conds() and
Item_subselect::accumulate_join_condition() with the new walk_join_list().

Reviewed By : Roy Lyseng <roy.lyseng@oracle.com>

Change-Id: I8702bc75cdb298ca302b0531b17af76acd22fd0d