Bug #30539757 NDB : NEW DOWNGRADE TESTS USING MGMD --INITIAL INCORRECTLY
Modify testUpgrade logic to correctly implement MGMD up/downgrade
as either :

 - Rolling restart, one at a time
   (Using cached binary configuration on disk)
 - Initial restart, all together on same version.
   (Parsing and regenerating configuration from config.ini)

This behaviour is under the control of the InitialMgmdRestart property
added in the fix for
  bug#30200250 DOWNGRADE TESTS FAIL DUE TO BINARY CONFIGURATION

There was also some preliminary work done to support MGMD initial restarts
in one testcase, in the changes pushed related to :

  WL#12233 MANUALLY START/STOP PROCESSES FROM TESTS

In this fix, a new method runChangeMgmds() encapsulates the work of
up/downgrading MGMDs.
Some tests explicitly require that 1 MGMD is upgraded at a time, and
so these cannot be run with --initial.  This constraint is checked.

The recently added test Downgrade_Mixed_MGMD_API_NDBD_WithMGMDInitialStart
attempts to perform a downgrade, with MGMDs restarted separately,
with --initial.
As this is not a supported scenario, the test cannot succeed reliably,
and so this test is removed.

Downgrades from versions with the test existing that do not already
fail on the rolling MGMD --initial starts, will fail as the postUpgrade
is missing in fixed versions.  These failures will clear away when
all *from* versions have this fix to testUpgrade.

Approved by : Varun Nagaraju <varun.nagaraju@oracle.com>