WL#5655 - InnoDB: Separate doublewrite file to ensure atomic writes.
The InnoDB doublewrite buffer lives in the system tablespace. Split the
doublewrite buffer to a separate file(s). This will  benefit users of SSD.
See notes in Bug#56283 for some details.

The pages of the legacy (v1) doublewrite buffer implementation will be
left as is and wil be reserved (but not used) when creating a new
instance. This is to make upgrade from 5.7 easy and maintain backward
compatibility with older tablespace formats. These pages and all
vestiges of the legacy doublewrite buffer will be removed later.

The contention on the double write mutex causes a bottleneck on our IO
throughput and increases the write latency.

RB:20522
Approved-by: Darshan M. N <darshan.m.n@oracle.com>
Approved-by: Mayank Prasad <mayank.prasad@oracle.com>
(cherry picked from commit f6903aac8f1ca88cbf394aa2540af6d91a95d2ae)