WL#13572: MEB: Add support for non-TTS partial restore
Description:
------------
MEB has support for non-TTS partial backup, backing up a selection of
InnoDB tables using TTS and partial restore of a backup taken using TTS.

This WL will implement support for non-TTS partial restore i.e one can
restore the selected tables from a non-TTS backup.

Implementation details:
-----------------------
It should be simple to filter only the selection of tables from a
full/partial backup and copy them to the target directories (data
directory and external data directories). But running apply-log could
apply the log to other existing tables in these directories as well
since apply-log works based on space-ids, not path names. If there is a
space-id match on destination server, that table would be updated. This
can invalidate the other tablespaces which are not part of partial
restore and crash the restore operation/server start-up.

In this WL, we won't load all the the existing tablespaces in the
beginning of the apply-log operation, but only the selected ones. This
will make sure that apply-log operation will not consider other
tablespaces and will ignore log records that are addressed to those
tablespaces.

Modified the code to construct the pattern specified with --include or
--include-tables or exclude-tables in options.cpp file. Throw an
exception and abort the operation if the pattern specified is invalid.

User visible changes:
---------------------
1. Option --include-tables is supported with restore operation of a
   full/partial backup and the tables matching the pattern specified
   with it will be restored.

2. Option --exclude-tables is supported with restore operation of a
   full/partial backup and the tables not matching the pattern
   specified with it will be restored.

RB: 23324
RB: 23592