BUG30639503: Ensure we keep all LDM threads busy
Approved-by: Ole John Aske <ole.john.aske@oracle.com>

The root fragment gets only a portion of the number of fragments to
scan. Thus the collective parallelism of all SPJ workers will be
the number of partitions in the root table if we ensure that each
SPJ worker will maintain a parallelism equal to the number of
fragments in the root table to handle.