Bug#30640463: ASSERTION "QEP_SHARED_OWNER..QS_CLEANUP"
The problem was that SHOW FIELDS/COLUMNS on temporary MERGE engine
tables with a subquery could lead to assert or server exit.

This was caused by a regression introduced in 8.0.16 by the patch
for Bug#28811287. The regression was that we did no longer take into
account that the table(s) underlying MERGE engine tables would
already be locked and opened when SHOW FIELDS/COLUMNS was
processed.

This patch fixes the problem by reintroducing code to handle this case.

Change-Id: Id0fa52c59bea67cf0701731594dd865efc444174