Bug#30664881 MAIN.HISTOGRAMS NOT REMOVING CREATED VIEW
main.histograms did not clean up properly; two views and a file
was left behind. Normally, this is caught by "check testcase",
but this is not always run (e.g., it is not run if the test has done
a server restart with non-default parameters). To ensure that missing
cleanup is caught in the future, do a server restart with default
parameters at the very end of the test.

Change-Id: I388ca6b81fc8632a18f10d514733fa3b1329b501