Bug #30655712: EMPTY RESULTSET FOR SUBQUERY WITH UNION AND TEXT COLUMN WITH CONSTANT EXPRESSION
In derived tables with unions known to be exactly one row (due to being a UNION
of a const table and zero or more known-zero expressions), we could materialize
the result into a temporary table that was never set as “not started”.
This would in turn cause the derived table to be incorrectly read as zero rows.

This was a regression in

  WL #13000: Iterator UNION [patch 12/16; direct materialization]

although I'm not sure exactly what change in there broke it. It's easy to fix
by simply marking the table as not started.

Change-Id: Ia128a7899962a81ad4ef470f33c9be9d89601b1a