WL#13377 Add support for hash outer, anti and semi join [2/8, --sorted_result]
Preparatory patch: Add --sorted_result to all tests that are
affected by upcoming patches.

Also, some of the tests are joins with a LIMIT, but without
any ORDER BY. Some of these will be executed using hash join
in one of the upcoming patches, meaning that the result set
may change. This patch adds an explicit ORDER BY to these
queries to make them deterministic.

Change-Id: I1f240fda229c4c673822f16ed9400b5aa6ca6e5a