WL#13377 Add support for hash outer, anti and semi join [4/8, refactor equi-join detection]
Preparatory patch: Since we now convert all BNL queries to use
hash join, we can simplify the code for determining whether
a condition can be used to execute hash join. We now only need to
classify a condition into two types; whether it contains _only_
equi-join conditions or not. For conditions that contains only
equi-join conditions, they are attached to the HashJoinIterator.
All other conditions are attached as filters after the join.

Change-Id: Ib759c9be0143bd75b8bec978bf7f3184d7649171