Bug #30473261: CONVERT THE INDEX SUBQUERY ENGINES INTO USING THE ITERATOR EXECUTOR [patch 1/10, case folding]
Fix broken case folding in subquery_sj.inc.

There was a regex in the test file that was probably intended to work around
differences between case-sensitive and case-insensitive platforms, but it was
way too broad, leading to “explAin” and so on. Fix the regex to only work on
the separate tokens “a” and “b”, not every letter a and b anywhere.

Change-Id: I760a0c8712df72e5a6a310aae869e3ba447ccc84