WL #13476: BKA outer/semi/anti join in iterator executor [patch 4/5, semijoin]
Implement semijoin for BKA. We implement this similarly to the existing
join buffering; keep a true/false “match flag” for each row to see
whether we've seen the inner row before, and if so, tell MRR to ignore the row
on future occasions. Unlike the join buffering, though, we don't keep the
flag in the row buffer; instead, we have a simple bit array. This saves
RAM, and keeps us from cluttering up the row buffer format with match flags.

Since ConnectJoins() is structured to do the right part of non-inner joins in a
recursive call (where we don't immediately know e.g. the outer tables),
this means we'll have to restructure the creation of MRR iterators a bit.
The MultiRangeRowIterator now gets a two-stage construction, where some of
the needed fields are not filled in until the parent BKAIterator is constructed.

Change-Id: I48ad55f1f494d628383f600535850800876da278