Bug #30417361: MAKE THE SORTING CODE INDEPENDENT OF QEP_TABS [patch 6/7, QEP_TAB from filesort]
Remove the last remnants of QEP_TABs from filesort; all we needed it for
now was some optimizer trace (which we replace with the table name, which
is not ideal, but better than just removing it) and whether we should free
buffers after sort or not (which can easily be decided on the outside
for the one case where it's needed). It remains in SortingIterator
until the next patch.

Change-Id: I5ba54592ed490a57db226fcb6f0cfc557dc30f12