Bug #30417361: MAKE THE SORTING CODE INDEPENDENT OF QEP_TABS [patch 5/7, gcol]
Unify the handling of scanning covering indexes containing virtual generated
columns.

The problem is that if A is (virtually) generated from B, we would normally
have both A and B in the read sets. But if we scan a covering index that
includes A, we would not have B available, and should not try to pack it
when we need to pack rows. Join buffering, hash join and filesort (the three
executor components that need to pack and unpack rows) each tackled
this problem more or less independently, which made for duplicated code,
and all three of them tried to switch back and forth. Instead, we set it up
once before join buffering, which is a lot simpler, and also increases
modularity.

There is a detail here in that occasionally, we may want to switch back to
table scans, and then B needs to go back into the read set so that it can
generate A. This only happens in two places (DynamicRangeIterator and
AlternativeIterator), which can take care of it locally.

Change-Id: I9ae333d0944eb4548cc77aa5d9970ac8b40d87e6