Bug #30556257: WEEDOUTITERATOR::READ HIT AN ASSERT DURING CHECK_UNIQUE_CONSTRAINT()
As part of its normal operation, duplicate weedout uses a table of row IDs it has
seen so far, with a unique index. However, if the key for that unique index becomes
too large (which can happen with very large row IDs), it revers to
deduplication by hash key instead (with a special non-unique index over the
hash field only), like other temporary tables. In this case, we would forget to
properly initialize said index, leading to an assertion failure in debug mode.

Change-Id: I1af1638d10918e073105fc1a3e7fb06035ea24e6