Bug #30664671: HASH JOINS : CRASH WITH COMPARING DIFFERENT CHARSET Bug #30664831: SEVERE PERFORMANCE REGRESSION JOINING ON MEDIUMTEXT COLUMNS!
When hash joining two string columns that are potentially very large
(in particular, blob columns with PAD SPACE collations), don't store the entire
sort key in the row, since that may be slow and require large amounts of
memory; instead, just store the (collation-aware) hash and add the equality
comparison for rechecking.

This is similar in spirit to what filesort does when faced with such a case;
filesort then truncates the key to “max_sort_length” (by default 1024 bytes).
However, the added comparison afterwards makes sure we can never get a wrong
answer, even in the face of 64-bit hash collisions.

Change-Id: I3572d3f13ee870f9d87b29533d5628ae69a36331