Bug#30566518 - SLAVE STOPS WITH INTERMITTENT ERROR 1396: ERROR                'OPERATION DROP USER FAILED FOR Bug#30324661 - INVALID GRANT COMMAND BREAK GTID BASED REPLICATION
Background
==========
Incident events are used to inform the slave that something out of
ordinary happened on the master that might cause the database to be in
an inconsistent state.

After bugfix for Bug#14598585, for any partially failed GRANT statements
we log an incident event to make the slave go down with a message
stating that manual reconciliation is needed.

Problem & Analysis
==================

1. Bug#30566518

Partially executing GRANT statements, sometimes do not write Incident
event (LOST_EVENTS) to the binary log.

Consider the below query that mixes both GLOBAL ACLs and DB_ACLs to a
non existing user.

eg: GRANT PROCESS ON `invalid_db`.* TO 'non_existing_user';

Since the user is not existing, the above query creates the user by
updating the mysql.user table. After the user is created, the query will
error out with ER_WRONG_USAGE (for mixing GLOBAL ACLs and DB_ACLs)

When this happens, neither the GRANT query nor the Incident event will
be recorded in the binary log (partial execution). In other words, the
cases of mixing different ACLs are not being handled properly causing
the master to diverge from the slave with an extra user.


2. Bug#30324661

After bugfix for Bug#19594845, we write the Incident event into the
stmt_cache, so that a GTID is generated prior to flushing the
stmt_cache. So, every incident events will have unique gtids when GTID
MODE=ON.

The above fix will work only when there is an active cache manager to
manage the stmt_cache and trx_cache. Since GTID generation is tightly
coupled with these caches, GTID will not be generated if there are no
such caches.

The above fix doesn't take into an account that, there will not be any
cache manager for the first query that is executed in the session.

As a result, if the first query in the session itself results in
generation of incident events, then that query will be directly logged
into binary log without any GTID, even when GTID_MODE=ON.

When slave is running with GTID_MODE=ON, it expects a valid GTID before
Incident event. Since GTID was not logged to the binary log on master,
slave assumes that the GTID to be anonymous and will error out with

slave> SHOW SLAVE STATUS\G
...
Last_Errno: 1782
Last_Error: Error executing incident event: '@@SESSION.GTID_NEXT cannot
            be set to ANONYMOUS when @@GLOBAL.GTID_MODE = ON.'


Fix
===
1. For Bug#30566518,

When GRANT query has a mix-up of different privileges, we mark those
transactions as partial transactions and we write Incident event to the
binary log so that slave applier thread stop with ER_SLAVE_INCIDENT.

2. For Bug#30324661,

While writing incident events, if we find that there is no cache manager
exists for the session, we create a cache manager to handle the
statement and transactional caches, so that a GTID is generated and is
written prior to flushing the stmt_cache. If there is any error while
creating such caches, then we trigger binlog_error_action if GTIDs are
enabled.

RB: 23543

Reviewed by: Harin Vadodaria <harin.vadodaria@oracle.com>
Reviewed by: Pedro Figueiredo <pedro.figueiredo@oracle.com>